+++ sdk/nodejs/neptune/subnetGroup.ts
@@ -62,7 +62,7 @@ export class SubnetGroup extends pulumi.CustomResource {
     /**
      * The description of the neptune subnet group. Defaults to "Managed by Pulumi".
      */
-    public readonly description!: pulumi.Output<string | undefined>;
+    public readonly description!: pulumi.Output<string>;
     /**
      * The name of the neptune subnet group. If omitted, this provider will assign a random, unique name.
      */
@@ -103,7 +103,7 @@ export class SubnetGroup extends pulumi.CustomResource {
             if (!args || args.subnetIds === undefined) {
                 throw new Error("Missing required property 'subnetIds'");
             }
-            inputs["description"] = args ? args.description : undefined;
+            inputs["description"] = (args ? args.description : undefined) || "Managed by Pulumi";
             inputs["name"] = args ? args.name : undefined;
             inputs["namePrefix"] = args ? args.namePrefix : undefined;
             inputs["subnetIds"] = args ? args.subnetIds : undefined;