+++ sdk/python/pulumi_aws/mediapackage/channel.py
@@ -68,6 +68,8 @@ class Channel(pulumi.CustomResource):
             if channel_id is None:
                 raise TypeError("Missing required property 'channel_id'")
             __props__['channel_id'] = channel_id
+            if description is None:
+                description = 'Managed by Pulumi'
             __props__['description'] = description
             __props__['tags'] = tags
             __props__['arn'] = None